put test in separate module
